Further precompiled header improvement

+++ Sources/Devices/Monitor.cpp
@@ -16,7 +16,7 @@ namespace acid
 		int32_t xpos;
 		int32_t ypos;
 		glfwGetMonitorPos(m_monitor, &xpos, &ypos);
-		return Vector2(xpos, ypos);
+		return Vector2(static_cast<float>(xpos), static_cast<float>(ypos));
 	}
 
 	Vector2 Monitor::GetDimensions() const
@@ -24,7 +24,7 @@ namespace acid
 		int32_t widthMM;
 		int32_t heightMM;
 		glfwGetMonitorPhysicalSize(m_monitor, &widthMM, &heightMM);
-		return Vector2(widthMM, heightMM);
+		return Vector2(static_cast<float>(widthMM), static_cast<float>(heightMM));
 	}
 
 	std::string Monitor::GetName() const
