+++ Sources/Particles/ParticleSystem.cpp
@@ -46,7 +46,7 @@ namespace acid
 
 			for (uint32_t i = 0; i < pastFactor; i++)
 			{
-				Particles::Get()->AddParticle(EmitParticle(*emitters[static_cast<uint32_t>(Maths::Random(0, emitters.size()))]));
+				Particles::Get()->AddParticle(EmitParticle(*emitters[static_cast<uint32_t>(Maths::Random(0.0f, static_cast<float>(emitters.size())))]));
 			}
 		}
 	}
@@ -157,7 +157,7 @@ namespace acid
 		velocity = velocity.Normalize();
 		velocity *= GenerateValue(m_averageSpeed, m_speedDeviation);
 
-		auto emitType = m_types.at(static_cast<uint32_t>(std::floor(Maths::Random(0, m_types.size()))));
+		auto emitType = m_types.at(static_cast<uint32_t>(std::floor(Maths::Random(0.0f, static_cast<float>(m_types.size())))));
 		float scale = GenerateValue(emitType->GetScale(), m_scaleDeviation);
 		float lifeLength = GenerateValue(emitType->GetLifeLength(), m_lifeDeviation);
 		float stageCycles = GenerateValue(emitType->GetStageCycles(), m_stageDeviation);
